pyproject.toml
[project]
name = "firengine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aioreactive>=0.20.0",
    "aiosql>=13.4",
    "ccxt>=4.4.65",
    "fastapi[standard]>=0.115.7",
    "ibis-framework[clickhouse,duckdb,polars]>=10.4.0",
    "keyring>=25.6.0",
    "kubernetes>=32.0.0",
    "nats-py>=2.10.0",
    "orjson>=3.10.16",
    "plotly>=6.0.1",
    "polars>=1.26.0",
    "pydantic-settings>=2.7.1",
    "pyiceberg>=0.9.0",
    "pyrsistent>=0.20.0",
    "python-json-logger>=3.3.0",
    "pyzmq>=26.2.1",
    "rich>=13.9.4",
    "shortuuid>=1.0.13",
    "structlog>=25.4.0",
    "typer>=0.15.1",
    "uvloop>=0.21.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.9.3",
]

[project.scripts]
fli = "firengine.cli.main:main"

[tool.uv]
package = true


[tool.ruff]
# Set as desired, typically 88 (black standard) or 100 (wide).
line-length = 120

[tool.ruff.lint]
# See: https://docs.astral.sh/ruff/rules/
select = [
    # Basic list from: https://docs.astral.sh/ruff/linter/#rule-selection
    "E",  # https://docs.astral.sh/ruff/rules/#error-e
    "F",  # https://docs.astral.sh/ruff/rules/#pyflakes-f
    "UP", # https://docs.astral.sh/ruff/rules/#pyupgrade-up
    "B",  # https://docs.astral.sh/ruff/rules/#flake8-bugbear-b
    "I",  # https://docs.astral.sh/ruff/rules/#isort-i
    # Other possibilities:
    # "D" # https://docs.astral.sh/ruff/rules/#pydocstyle-d
    # "Q" # https://docs.astral.sh/ruff/rules/#flake8-quotes-q
    # "COM" # https://docs.astral.sh/ruff/rules/#flake8-commas-com
    # "SIM", # https://docs.astral.sh/ruff/rules/#flake8-simplify-sim

]
ignore = [
    "E501", # https://docs.astral.sh/ruff/rules/line-too-long/
    "E402", # https://docs.astral.sh/ruff/rules/module-import-not-at-top-of-file/
    "E731", # https://docs.astral.sh/ruff/rules/lambda-assignment/
    # We use both ruff formatter and linter so some rules should always be disabled.
    # See: https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules
    "W191",   # https://docs.astral.sh/ruff/rules/tab-indentation/
    "E111",   # https://docs.astral.sh/ruff/rules/indentation-with-invalid-multiple/
    "E114",   # https://docs.astral.sh/ruff/rules/indentation-with-invalid-multiple-comment/
    "E117",   # https://docs.astral.sh/ruff/rules/over-indented/
    "D206",   # https://docs.astral.sh/ruff/rules/docstring-tab-indentation/
    "D300",   # https://docs.astral.sh/ruff/rules/triple-single-quotes/
    "Q000",   # https://docs.astral.sh/ruff/rules/bad-quotes-inline-string/
    "Q001",   # https://docs.astral.sh/ruff/rules/bad-quotes-multiline-string/
    "Q002",   # https://docs.astral.sh/ruff/rules/bad-quotes-docstring/
    "Q003",   # https://docs.astral.sh/ruff/rules/avoidable-escaped-quote/
    "COM812", # https://docs.astral.sh/ruff/rules/missing-trailing-comma/
    "COM819", # https://docs.astral.sh/ruff/rules/prohibited-trailing-comma/
    "ISC002", # https://docs.astral.sh/ruff/rules/multi-line-implicit-string-concatenation/
]
